RefineSegFormer/requirements.txt
torch==1.13.0+cu117
torchvision==0.14.0+cu117
tensorboard==2.14.0
scipy==1.9.3
numpy==1.24.4
opencv_python==4.8.1.78
tqdm==4.67.1
Pillow==10.4.0